+++ changelog.md
@@ -1,3 +1,9 @@
+0.0.2.0 2026-07-17
+==================
+- update sway-binds & sway-show-bindings to handle --locked, and
+  [ Fn+p, xf86launch2 (Fn+F10), xf86favorites (Fn+F12) ], and use 𐌎 for
+  windoze key
+
 0.0.1.0 2026-07-08
 ==================
 - upgrade to nixos-26.05